Fixed up solid start
