+++ packages/app-solid-start/src/components/ClientSideRenderedDetail.tsx
@@ -1,0 +1,7 @@
+import { useParams } from '@solidjs/router'
+
+export default function ClientSideRenderedDetail() {
+  const params = useParams()
+
+  return <p id="detail-id">{params.id}</p>
+}
