+++ packages/app-solid-start/src/routes/client-side-rendered/index.tsx
@@ -1,31 +1,9 @@
-import { For } from 'solid-js'
-import { A } from '@solidjs/router'
+import { clientOnly } from '@solidjs/start'
 
-export const ssr = false
-
-type Entry = { id: string; name: string }
-
-const entries: Entry[] = Array.from({ length: 1000 }, () => ({
-  id: crypto.randomUUID(),
-  name: crypto.randomUUID(),
-}))
+const ClientSideRenderedTable = clientOnly(
+  () => import('../../components/ClientSideRenderedTable'),
+)
 
 export default function ClientSideRenderedPage() {
-  return (
-    <table>
-      <tbody>
-        <For each={entries}>
-          {(entry) => (
-            <tr>
-              <td>{entry.id}</td>
-              <td>{entry.name}</td>
-              <td>
-                <A href={`/client-side-rendered/${entry.id}`}>View →</A>
-              </td>
-            </tr>
-          )}
-        </For>
-      </tbody>
-    </table>
-  )
+  return <ClientSideRenderedTable />
 }
